Add maintained GitHub tracker adapter
