+++ src/tracker/factory.ts
@@ -1,11 +1,14 @@
 import type { TrackerConfig } from "../domain/workflow.js";
 import type { Logger } from "../observability/logger.js";
 import type { Tracker } from "./service.js";
+import { GitHubTracker } from "./github.js";
 import { GitHubBootstrapTracker } from "./github-bootstrap.js";
 import { LinearTracker } from "./linear.js";
 
 export function createTracker(config: TrackerConfig, logger: Logger): Tracker {
   switch (config.kind) {
+    case "github":
+      return new GitHubTracker(config, logger);
     case "github-bootstrap":
       return new GitHubBootstrapTracker(config, logger);
     case "linear":
